# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Biblioteca API",
        "description": "Documentación de la API de la biblioteca",
        "contact": {
            "email": "soporte@bibliotecaapi.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/loans": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "List all loans",
                "operationId": "53fdef403307586122900dc3b1d07368",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Loan"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Loans"
                ],
                "summary": "Create a new loan",
                "operationId": "92bd5c653a9a8345b4f161d5ba3bbd71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "book_id",
                                    "loan_date"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "example": 11
                                    },
                                    "loan_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-01"
                                    },
                                    "return_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-15",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Loan created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Loan"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/loans/{loan}": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "Get a loan by ID",
                "operationId": "39a4e10f305b5be4c3e4eba31a94777a",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "description": "Loan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Loan"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Loan not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Loans"
                ],
                "summary": "Update a loan",
                "operationId": "d1a9dfd54e508debb4769bc440e372f8",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "description": "Loan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "example": 11
                                    },
                                    "loan_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-01"
                                    },
                                    "return_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-15",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Loan updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Loan"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Loan not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Loans"
                ],
                "summary": "Delete a loan",
                "operationId": "6d80bd9e85d71ce0f714269e753f85fb",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "description": "Loan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Loan deleted"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Loan not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Loan": {
                "title": "Loan",
                "required": [
                    "id",
                    "user_id",
                    "book_id",
                    "loan_date"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "book_id": {
                        "type": "integer",
                        "example": 11
                    },
                    "loan_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-01"
                    },
                    "return_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-15",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}